+++ src/traits.rs
@@ -1,8 +1,12 @@
-use std::{error::Error, fmt::Debug, hash::Hash, sync::Arc};
+use std::{fmt::Debug, sync::Arc};
 
 use async_trait::async_trait;
+use tokio::{
+    fs::{self, OpenOptions},
+    io::{AsyncRead, AsyncWrite},
+};
 
-use crate::Path;
+use crate::{Error, Path};
 
 /// A datatype that can be used as the access key for cached items.
 ///
@@ -11,10 +15,12 @@ use crate::Path;
 /// UUIDv4 for most use cases.
 pub trait Key
 where
-    Self: Clone + Debug + Eq + Hash,
+    Self: Debug + 'static,
 {
     /// Generate a new, unique key.
-    fn new() -> Self;
+    fn new() -> Self
+    where
+        Self: Sized;
 
     /// Convert this key to a filename used for flushing to disk.
     fn as_filename(&self) -> String;
@@ -37,20 +43,54 @@ pub trait AsyncFileRepr
 where
     Self: Sized,
 {
-    type Err: Error;
+    type Err: std::error::Error;
 
-    /// Load the data structure from disk asynchronously.
+    /// Load (deserialise) the data structure into memory asynchronously.
     ///
-    /// If you wish to perform non-trivial deserialisation in this function,
+    /// If you wish to perform non-trivial computation/conversion in this function,
     /// you should spawn a blocking task with your async runtime.
-    async fn load(path: impl AsRef<Path> + Send) -> Result<Self, Self::Err>;
+    async fn load<R>(reader: R) -> Result<Self, Self::Err>
+    where
+        R: Send + Unpin + AsyncRead;
 
-    /// Flush the data structure to disk asynchronously.
+    /// Flush (serialise) the data structure from memory asynchronously.
     ///
-    /// If you wish to perform non-trivial serialisation in this function,
+    /// If you wish to perform non-trivial computation/conversion in this function,
     /// you should spawn a blocking task with your async runtime.
-    async fn flush(self: &Arc<Self>, path: impl AsRef<Path> + Send) -> Result<(), Self::Err>;
+    async fn flush<W>(self: &Arc<Self>, writer: W) -> Result<(), Self::Err>
+    where
+        W: Send + Unpin + AsyncWrite;
 
-    /// Delete the data structure from disk asynchronously.
-    async fn delete(path: impl AsRef<Path> + Send) -> Result<(), Self::Err>;
+    /// Load (deserialise) the data structure from disk.
+    async fn load_from_disk(path: impl AsRef<Path> + Send) -> Result<Self, Error<Self::Err>> {
+        let file = OpenOptions::new().read(true).open(path).await?;
+        let data = Self::load(file).await.map_err(Error::Serde)?;
+        Ok(data)
+    }
+
+    /// Flush (serialise) the data structure to disk.
+    async fn flush_to_disk(
+        self: &Arc<Self>,
+        path: impl AsRef<Path> + Send,
+    ) -> Result<(), Error<Self::Err>>
+    where
+        Self: Send,
+    {
+        let file = OpenOptions::new()
+            .write(true)
+            .create(true)
+            .truncate(true)
+            .open(path)
+            .await?;
+        self.flush(file).await.map_err(Error::Serde)?;
+        Ok(())
+    }
+
+    /// Delete the data structure from disk.
+    ///
+    /// Override this method if you wish to perform extra cleanup before deletion.
+    async fn delete(path: impl AsRef<Path> + Send) -> Result<(), Error<Self::Err>> {
+        fs::remove_file(path).await?;
+        Ok(())
+    }
 }