+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.1.0
 
+- Reworked traits
 - Copied from https://github.com/cyqsimon/file-lfu